+++ tests/test_link_order1.v
@@ -24,4 +24,4 @@ Elpi Accumulate lp:{{
 }}.
 Elpi Typecheck.
 Elpi add 10.
-Elpi Print foo "tests/test_link_order1".
+Elpi Print foo "elpi.tests/test_link_order1".