+++ tests/test_link_order5.v
@@ -25,4 +25,4 @@ Elpi Accumulate lp:{{
 Elpi Typecheck.
 Elpi add 5.
 Elpi add 5.
-Elpi Print foo "tests/test_link_order5".
+Elpi Print foo "elpi.tests/test_link_order5".