+++ tests/test_link_order9.v
@@ -2,5 +2,5 @@ From elpi Require Import elpi.
 From elpi.tests Require Import test_link_order1.
 From elpi.tests Require Import test_link_order1.
 
-Elpi Print foo "tests/test_link_order9".
+Elpi Print foo "elpi.tests/test_link_order9".